updating structure and style #2
