+++ episode_26_whereToNow.qmd
@@ -5,7 +5,7 @@ toc: true
 
 You have now seen the basics of using Quarto for creating documents. Quarto is a relatively new product from posit (the group behind RStudio) and there are regular, exciting updates on the project. While we have only given a brief introduction here, there are many excellent examples and resources available online to take you further with Quarto.
 
-The key resource is the [Quarto webiste](https://quarto.org/). If you have completed everything in this workshop then you may want to skip to the Guide tab, or use the Get Started tab to introduce you to using Quarto in a different tool (*e.g.,* VSCode instead of RStudio). 
+The key resource is the [Quarto website](https://quarto.org/). If you have completed everything in this workshop then you may want to skip to the Guide tab, or use the Get Started tab to introduce you to using Quarto in a different tool (*e.g.,* VSCode instead of RStudio). 
 
 The [posit::conf(2023)](https://www.youtube.com/playlist?list=PL9HYL-VRX0oRFZslRGHwHuwea7SvAATHp) YouTube playlist has recordings from 14 talks (!!!) that mention Quarto, and in 2024 the conference organisers [created a specific Quarto playlist](https://www.youtube.com/playlist?list=PLNUVZZ6hfXX1Y4Is-SbbMF_HutRDJBwiO). If nothing else, this should show you how important the community things Quarto is for bioinformatics. Key examples I can recommend include talks on automated reporting and scrollytelling.
